requirements.txt
albumentations==1.3.0
comet_ml==3.32.6
coremltools==6.3.0
export==0.2.0
Flask==2.2.2
Flask_Cors==3.0.10
GitPython==3.1.31
ipython==8.12.0
matplotlib==3.5.1
mss==7.0.1
numpy==1.21.5
onnxruntime==1.14.1
opencv_python==4.7.0.72
openvino==2022.3.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
PyYAML==6.0
requests==2.25.1
scipy==1.10.1
seaborn==0.12.2
setuptools==65.6.3
tensorflow==2.12.0
tensorrt==8.6.0
tflite_runtime==2.11.0
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
train==0.0.5
tritonclient==2.32.0
Werkzeug==2.2.3
